mem0-agent/requirements.txt
pywin32==308; sys_platform == "win32"

aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
argcomplete==3.6.0
attrs==25.1.0
backoff==2.2.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cohere==5.14.0
colorama==0.4.6
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
eval_type_backport==0.2.2
executing==2.2.0
fastapi==0.115.11
fastavro==1.10.0
filelock==3.17.0
flupy==1.2.1
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.69.0
gotrue==2.11.4
greenlet==3.1.1
griffe==1.6.0
groq==0.18.0
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.2
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.6
jiter==0.8.2
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
logfire==3.7.1
logfire-api==3.7.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mem0ai==0.1.65
mistralai==1.5.0
monotonic==1.6
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.29.1
numpy==2.2.3
openai==1.65.4
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
packaging==24.2
pandas==2.2.3
pgvector==0.3.6
pillow==11.1.0
portalocker==2.10.1
postgrest==0.19.3
posthog==3.18.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-ai==0.0.35
pydantic-ai-slim==0.0.35
pydantic-graph==0.0.35
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.13.3
realtime==2.4.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
setuptools==75.8.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
storage3==0.11.3
streamlit==1.43.0
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tenacity==9.0.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
types-requests==2.32.0.20250306
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vecs==0.4.5
watchdog==6.0.0
wcwidth==0.2.13
websockets==14.2
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
